# contrib/internal/httptrace: improve performance of span starts