--- a/contrib/internal/httptrace/httptrace.go
+++ b/contrib/internal/httptrace/httptrace.go
@@ -24,7 +24,7 @@ import (
 
 var (
 	cfg      = newConfig()
-	serverOp = namingschema.NewHTTPServerOp()
+	serverOp = namingschema.NewHTTPServerOp().GetName()
 )
 
 // StartRequestSpan starts an HTTP request span with the standard list of HTTP request span tags (http.method, http.url,
@@ -38,13 +38,6 @@ func StartRequestSpan(r *http.Request, opts ...ddtrace.StartSpanOption) (tracer.
 		ipTags, _ = httpsec.ClientIPTags(r.Header, true, r.RemoteAddr)
 	}
 	nopts := make([]ddtrace.StartSpanOption, 0, len(opts)+7+len(ipTags))
-	// nopts = append(nopts,
-	// 	tracer.SpanType(ext.SpanTypeWeb),
-	// 	tracer.Tag(ext.HTTPMethod, r.Method),
-	// 	tracer.Tag(ext.HTTPURL, urlFromRequest(r)),
-	// 	tracer.Tag(ext.HTTPUserAgent, r.UserAgent()),
-	// 	tracer.Measured())
-
 	nopts = append(nopts,
 		func(cfg *ddtrace.StartSpanConfig) {
 			if cfg.Tags == nil {
@@ -66,7 +59,7 @@ func StartRequestSpan(r *http.Request, opts ...ddtrace.StartSpanOption) (tracer.
 			}
 		})
 	nopts = append(nopts, opts...)
-	return tracer.StartSpanFromContext(r.Context(), serverOp.Name(), nopts...)
+	return tracer.StartSpanFromContext(r.Context(), serverOp, nopts...)
 }
 
 // FinishRequestSpan finishes the given HTTP request span and sets the expected response-related tags such as the status