@knusbaum 2023-07-19
```
$ benchstat old.bench new.bench 
name                 old time/op    new time/op    delta
StartRequestSpan-16     788ns ±13%     160ns ± 6%  -79.64%  (p=0.000 n=10+10)

name                 old alloc/op   new alloc/op   delta
StartRequestSpan-16      768B ± 0%      224B ± 0%  -70.83%  (p=0.000 n=10+10)

name                 old allocs/op  new allocs/op  delta
StartRequestSpan-16      21.0 ± 0%       5.0 ± 0%  -76.19%  (p=0.000 n=10+10)
```