@knusbaum 2023-07-19
New benchstat output:
```
$ benchstat old.bench new.bench 
name                 old time/op    new time/op    delta
StartRequestSpan-16    1.08µs ±13%    0.35µs ± 3%  -67.92%  (p=0.000 n=10+10)

name                 old alloc/op   new alloc/op   delta
StartRequestSpan-16      768B ± 0%      264B ± 0%  -65.62%  (p=0.000 n=10+10)

name                 old allocs/op  new allocs/op  delta
StartRequestSpan-16      21.0 ± 0%      10.0 ± 0%  -52.38%  (p=0.000 n=10+10)
```